feat: enhance internationalization support and validate language/theme settings
- Added language and theme validation in the loader function to ensure only supported values are used.
- Refactored language switching to utilize the new Language type for better type safety.
- Moved SUPPORTED_LANGUAGES definition to the i18n config for centralized management.
- Updated relevant components and hooks to align with the new language handling structure.
